refactor: `Item` , `itemDto` 리팩토링 (#69)
* 클래스명 수정 `item` -> `Item`
* 중간 연관 테이블을 따로 뒀기에 더 이상 `Item` 은 `Article` 을 가지지 않기로 함
* 삭제됐더라도 다시 추가됐을 때 삭제 상태를 변경하기 위한 코드 추가
